refactor(rust): move StaticArray to polars-arrow
