+++ crates/polars-arrow/src/array/mod.rs
@@ -668,6 +668,8 @@ mod map;
 mod null;
 mod primitive;
 pub mod specification;
+mod static_array;
+mod static_array_collect;
 mod struct_;
 mod total_ord;
 mod union;
@@ -697,6 +699,8 @@ pub use map::MapArray;
 pub use null::{MutableNullArray, NullArray};
 use polars_error::PolarsResult;
 pub use primitive::*;
+pub use static_array::{ParameterFreeDtypeStaticArray, StaticArray};
+pub use static_array_collect::{ArrayCollectIterExt, ArrayFromIter, ArrayFromIterDtype};
 pub use struct_::{MutableStructArray, StructArray};
 pub use union::UnionArray;
 pub use utf8::{MutableUtf8Array, MutableUtf8ValuesArray, Utf8Array, Utf8ValuesIter};
